refactor: move related cobbler tests to standalone classes
